+++ source/redux/HttpClient.js
@@ -1,35 +1,35 @@
 import HttpClient from '../HttpClient'
 
-export default function createHttpClient(settings, get_store, protected_cookie_value, options = {})
+export default function createHttpClient(settings, get_store, options = {})
 {
 	let on_before_send
 	let catch_to_retry
 	let get_access_token
 
-	// Add `store` helper to `http.request`
-	if (settings.http.request)
-	{
-		on_before_send = (request) =>
-		{
-			// If using Redux, then add `store` as a parameter
-			// for `http_client` customization function
-			settings.http.request(request,
-			{
-				store: get_store()
-			})
-		}
-	}
+	// // Add `getState` to `http.request` parameters.
+	// if (settings.http.request)
+	// {
+	// 	on_before_send = (request) =>
+	// 	{
+	// 		// If using Redux, then add `store` as a parameter
+	// 		// for `http_client` customization function
+	// 		settings.http.onBeforeSend(request,
+	// 		{
+	// 			getState: get_store().getState
+	// 		})
+	// 	}
+	// }
 
 	// Add `store` and `http` helpers to `http.catch`
 	if (settings.http.catch)
 	{
 		catch_to_retry = (error, retryCount, helpers) =>
 		{
-			return settings.http.catch(error, retryCount,
-			{
-				...helpers,
-				store: get_store()
-			})
+			return settings.http.catch(error, retryCount, helpers)
+			// {
+			// 	...helpers,
+			// 	store: get_store()
+			// })
 		}
 	}
 
@@ -55,8 +55,6 @@ export default function createHttpClient(settings, get_store, protected_cookie_v
 		allow_absolute_urls         : settings.http.allowAbsoluteURLs,
 		parseDates                  : settings.parseDates,
 		authentication_token_header : settings.authentication.header,
-		protected_cookie            : settings.authentication.protectedCookie,
-		protected_cookie_value,
 		...options
 	})
 }